SHIN-EKI
This is the second map I have made for Bad Business. It was originally an idea by Ethan who later dropped it. I picked up his idea, being heavily inspired by the real-life counterpart SHINJUKU-EKI and created SHIN-EKI. Map Cinematic - https://youtu.be/NFkKd9GsJ7s

It's 2010. A doomsday cult hijacks a train and tries driving the train into the city, but their train is stopped, and they are left stranded in SHIN-EKI. With the SAT entering the station, the doomsday cult heads up into the mall and prepares to get to the control room to restart the train... 

2 and a half months of researching and map building, less than 2 weeks of scripting and filming. Thank you everyone who participated to make this trailer come to life.