NUKETOWN 2025
A proper Nuketown map on Roblox, specifically made for Bad Business. I no longer have the opinions in my "Cinematic" video for it, so it has been taken down.

Made up of 3k+ parts and only 2 months of building. I gave up on it 5 months later. The background is unfinished, but the map is completely playable and has only 3 minor clipping issues which I can no longer fix due to losing the original map file.